tested vnpay (via postman)

+++ SpaServiceBE/Repositories/Entities/Vnpay/PaymentInformationModel.cs
@@ -9,7 +9,7 @@ namespace Repositories.Entities.Vnpay
     public class PaymentInformationModel
     {
         public string OrderType { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string OrderDescription { get; set; }
         public string Name { get; set; }
